[MLS] Rename RegisterUserForm to UserForm to use it in another context
